+++ example/images/failing_app/image.rb
@@ -1,0 +1,4 @@
+KuberKit
+  .define_image(:failing_app)
+  .registry(:default)
+  .depends_on(:ruby, :app_sources)