+++ example/services/failing_app.rb
@@ -1,0 +1,6 @@
+KuberKit
+  .define_service(:failing_app)
+  .depends_on(:env_file)
+  .template(:service)
+  .images(:failing_app)
+  .tags("app", "minimal")